+++ addon/components/frost-list-item-first-column.js
@@ -1,5 +1,5 @@
 import Ember from 'ember'
-import layout from '../templates/components/frost-list-item-first-column'
+import layout from '../templates/frost-list-item-first-column'
 
 export default Ember.Component.extend({
   layout
