+++ addon/templates/frost-list-item-binding.hbs
