+++ addon/templates/frost-list-item-first-column.hbs
